neteja i refresh i ack positions paperera i Walls

+++ UDP/UDP_Game_Libreria/GameLib/GlobalValues.cpp
@@ -1,16 +1,17 @@
 #include "GlobalValues.h"
 
-Position PixelToCell(int8_t _x, int8_t _y)
+Position PixelToCell(int16_t _x, int16_t _y)
 {
-	int8_t xCell = _x / SIZE_CELL;
-	int8_t yCell = _y / SIZE_CELL;
+	int16_t xCell = _x / SIZE_CELL;
+	int16_t yCell = _y / SIZE_CELL;
 	return 	Position{ xCell, yCell };
 }
 
-Position CellToPixel(int8_t _x, int8_t _y)
+Position CellToPixel(int16_t _x, int16_t _y)
 {
 	return Position{ _x * SIZE_CELL, _y * SIZE_CELL };; //convert to pixels
 }
+
 
 sf::Packet& operator <<(sf::Packet& Packet, const Position& pos)
 {
